More changes to compile on macbook. static link fails
More changes to avoid compile errors on macbook.
The zlib Makefile tries to do a static & dynamic build. The
dyn build isn't needed and gets link errors so I did:
cd zlib*; ./configure --static
to just build the .a lib (which is all I need).
These changes let oppat compile but the static link fails.
Apparently it is a std problem on OS X.
More work on this eventually.

+++ src/prf_read_data.cpp
@@ -25,6 +25,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
+#include <vector>
 #include <cstddef>
 #include <inttypes.h>
 #include <algorithm>
@@ -737,7 +738,7 @@ static int prf_decode_perf_record(const long pos_rec, uint64_t typ, char *rec, i
 			int exp_identifier = (prf_obj.def_sample_flags & PERF_SAMPLE_IDENTIFIER ? 1 : 0);
 			if (!exp_id && !exp_identifier && prf_obj.events.size() != 1) {
 				printf("ummm... got no IDs in the prf event list but event list size is(%" PRIu64 ") != 1. ck yer logic dude. Bye at %s %d\n",
-					prf_obj.events.size(), __FILE__, __LINE__);
+					(uint64_t)(prf_obj.events.size()), __FILE__, __LINE__);
 				exit(1);
 			}
 			if (prf_obj.events.size() == 1) {
